superposewf object, should work fine now except for the pgradient
